+++ source/objects/Texture.cpp
@@ -74,7 +74,7 @@ void Texture::createTextureImage(const VkCommandPool& commandPool, const VkQueue
 
   void* data;
   vkMapMemory(device, stagingBufferMemory, 0, imageSize, 0, &data);
-  memcpy(data, pixels, static_cast<size_t>(imageSize));
+  memcpy(data, pixels, imageSize);
   vkUnmapMemory(device, stagingBufferMemory);
 
   stbi_image_free(pixels);
